Extracted functionality to base class
