+++ JHotDraw/src/main/java/org/jhotdraw/draw/action/ButtonFactory.java
@@ -13,6 +13,8 @@
  */
 package org.jhotdraw.draw.action;
 
+import org.jhotdraw.draw.action.arrange.BringToFrontAction;
+import org.jhotdraw.draw.action.arrange.SendToBackAction;
 import dk.sdu.mmmi.featuretracer.lib.FeatureEntryPoint;
 import org.jhotdraw.gui.JPopupButton;
 import org.jhotdraw.util.*;
