+++ JHotDraw/src/main/java/org/jhotdraw/draw/action/arrange/SendToBackAction.java
@@ -1,5 +1,5 @@
 /*
- * @(#)BringToFrontAction.java  2.0  2008-05-30
+ * @(#)SendToBackAction.java  2.0  2008-05-30
  *
  * Copyright (c) 2003-2008 by the original authors of JHotDraw
  * and all its contributors.
@@ -11,27 +11,30 @@
  * accordance with the license agreement you entered into with  
  * the copyright holders. For details see accompanying license terms. 
  */
-package org.jhotdraw.draw.action;
+package org.jhotdraw.draw.action.arrange;
 
+import dk.sdu.mmmi.featuretracer.lib.FeatureEntryPoint;
+import org.jhotdraw.app.JHotDrawFeatures;
 import org.jhotdraw.draw.*;
 
 /**
- * ToFrontAction.
+ * SendToBackAction.
  *
  * @author Werner Randelshofer
- * @version 2.0 2008-05-30 Renamed from MoveToFrontAction to BringToFrontAction
- * for consistency with the API of Drawing.
+ * @version 2.0 2008-05-30 Renamed from MoveToBackAction to SendToBackAction for
+ * consistency with the API of Drawing.
  * <br>1.0 24. November 2003 Created.
  */
-public class BringToFrontAction extends SendToBackAction {
+public class SendToBackAction extends ArrangeAction {
 
-    public static String ID = "edit.bringToFront";
+    public final static String ID = "edit.sendToBack";
 
     /**
      * Creates a new instance.
      */
-    public BringToFrontAction(DrawingEditor editor) {
-        super(editor, ArrangeLayer.FRONT);
+    @FeatureEntryPoint(JHotDrawFeatures.ARRANGE)
+    public SendToBackAction(DrawingEditor editor) {
+        super(editor, ID, ArrangeLayer.BACK);
         labels.configureAction(this, ID);
     }
 }
