+++ JHotDraw/src/main/java/org/jhotdraw/samples/svg/SVGApplicationModel.java
@@ -13,6 +13,8 @@
  */
 package org.jhotdraw.samples.svg;
 
+import org.jhotdraw.draw.action.arrange.BringToFrontAction;
+import org.jhotdraw.draw.action.arrange.SendToBackAction;
 import dk.sdu.mmmi.featuretracer.lib.FeatureEntryPoint;
 import org.jhotdraw.app.Application;
 import org.jhotdraw.app.DefaultApplicationModel;
