+++ JHotDraw/src/main/java/org/jhotdraw/samples/svg/gui/ArrangeToolBar.java
@@ -13,6 +13,8 @@
  */
 package org.jhotdraw.samples.svg.gui;
 
+import org.jhotdraw.draw.action.arrange.BringToFrontAction;
+import org.jhotdraw.draw.action.arrange.SendToBackAction;
 import java.beans.*;
 import java.util.prefs.*;
 import javax.swing.border.*;
